requirements.txt
django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
celery[redis]==5.2.7
psycopg2-binary==2.9.5
Pillow==9.3.0